adds 64 bitness to calabash.framework

+++ calabash/Classes/FranklyServer/Operations/LPQueryAllOperation.m
@@ -134,7 +134,7 @@
       case 'c':[invocation getReturnValue:(void **) &charValue];
         return [NSString stringWithFormat:@"%d", charValue];
       case '{': {
-        unsigned int length = [[invocation methodSignature] methodReturnLength];
+        NSUInteger length = [[invocation methodSignature] methodReturnLength];
         void *buffer = (void *) malloc(length);
         [invocation getReturnValue:buffer];
         NSValue *value = [[[NSValue alloc] initWithBytes:buffer objCType:type]
